requirements.txt
asttokens==2.0.8
backcall==0.2.0
bson==0.5.10
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.0.0
Flask==2.2.2
fpdf==1.7.2
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
ipdb==0.13.9
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.23.3
openpyxl==3.1.2
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
Pygments==2.13.0
pypdf2==3.0.1
pyspark==3.2.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.0
toml==0.10.2
traitlets==5.3.0
typing-extensions==4.7.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
XlsxWriter==3.1.2
zipp==3.8.1
